// projects/game-logic/src/lib/features/effects/commons/effect-payload-collector/effect-payload.interface.ts
export interface IPayloadDefinition {
  effect: IEffect;
  preparationSteps?: ICollectableDataDefinition[];
  gatheringSteps?: ICollectableDataDefinition[];
  amountOfTargets?: number;
  caster: IEffectCaster;
  nestedDefinitionFactory?: (prep: ICollectableData) => IPayloadDefinition
}

export interface ICollectableDataDefinitionBase {
  effectId?: string;
  dataName: GatheringStepDataName;
  requireUniqueness: boolean;
  autoCollect?: boolean;
  incorporatePayloadDefinitionForSelectedEffect?: boolean;
}

export interface ICollectableData {
  index: number;
  effect: IEffect;
  steps: ICollectableDataStep[];
  isCompleted: boolean;
  payloadDefinition: IPayloadDefinition;
}

export interface ICollectableDataStepBase {
  effectId: string;
  collectedDataIndex: number;
  dataName: GatheringStepDataName;
  prev?: ICollectableDataStep[];
  attemptWasMade: boolean;
}


export type ICollectableDataStep =
  IOriginCollectableDataStep |
  ISourceActorCollectableDataStep |
  IActorCollectableDataStep |
  IFieldCollectableDataStep |
  IEffectCollectableDataStep |
  IRotationCollectableDataStep;
// ...
export type ICollectableDataDefinition =
  IOriginCollectableDataDefinition |
  ISourceActorCollectableDataDefinition |
  IActorCollectableDataDefinition |
  IFieldCollectableDataDefinition |
  IEffectCollectableDataDefinition |
  IRotationCollectableDataDefinition;


export interface IOriginCollectableDataDefinition extends ICollectableDataDefinitionBase {
  dataName: GatheringStepDataName.Origin;
  possibleOriginsResolver?: (prev: ICollectableDataStep[]) => IBoardSelectorOrigin[];
  possibleOrigins?: IBoardSelectorOrigin[];
  payload?: IBoardSelectorOrigin;
  prev?: ICollectableDataStep[];
  initialPayload?: IBoardSelectorOrigin;
  initialPayloadResolver?: (prev: ICollectableDataStep[]) => IBoardSelectorOrigin | undefined;
}

export interface ISourceActorCollectableDataDefinition extends ICollectableDataDefinitionBase {
  dataName: GatheringStepDataName.SourceActor;
  possibleSourceActorIdsResolver?: (prev: ICollectableDataStep[]) => string[];
  possibleSourceActorIds?: string[];
  payload?: string;
  prev?: ICollectableDataStep[];
  initialPayload?: string;
  initialPayloadResolver?: (prev: ICollectableDataStep[]) => string | undefined;
}

export interface IActorCollectableDataDefinition extends ICollectableDataDefinitionBase {
  dataName: GatheringStepDataName.Actor;
  possibleActorsResolver?: (prev: ICollectableDataStep[]) => IActor[];
  possibleActors?: IActor[];
  payload?: IActor;
  prev?: ICollectableDataStep[];
  initialPayload?: IActor;
  initialPayloadResolver?: (prev: ICollectableDataStep[]) => IActor | undefined;
}

export interface IFieldCollectableDataDefinition extends ICollectableDataDefinitionBase {
  dataName: GatheringStepDataName.Field;  
  possibleFieldsResolver?: (prev: ICollectableDataStep[]) => IField[];
  possibleFields?: IField[];
  payload?: IField;
  prev?: ICollectableDataStep[];
  initialPayload?: IField;
  initialPayloadResolver?: (prev: ICollectableDataStep[]) => IField | undefined;
}

export interface IEffectCollectableDataDefinition extends ICollectableDataDefinitionBase {
  dataName: GatheringStepDataName.Effect;  
  possibleEffectsResolver?: (prev: ICollectableDataStep[]) => IEffect[];
  possibleEffects?: IEffect[];
  payload?: IEffect;
  prev?: ICollectableDataStep[];
  initialPayload?: IEffect;
  initialPayloadResolver?: (prev: ICollectableDataStep[]) => IEffect | undefined;
}

export interface IRotationCollectableDataDefinition extends ICollectableDataDefinitionBase {
  dataName: GatheringStepDataName.Rotation
  payload?: number;
  prev?: ICollectableDataStep[];
  initialPayload?: number;
  initialPayloadResolver?: (prev: ICollectableDataStep[]) => number | undefined;
}